fix(attack): cast semantic-target inputs to model.dtype (bf16 vision backbone)
build_semantic_target passed the processor's float32 pixel_values to OpenVLA's bf16
vision backbone -> RuntimeError 'Input type (float) and bias type (BFloat16) should be
the same' at model.generate. The torch-free [VERIFY on box] path was invisible to the
fake-model unit tests (dtype-agnostic). Fix casts the batch to model.dtype, matching
openvla_utils.get_vla_action + gcg_openvla; HF BatchFeature.to casts only float tensors,
leaving input_ids int. Box-verified: generate returns 7 action-range ids. + regression
test (204 attack tests green, ruff clean).

Records the completed LIBERO_Object benign baseline (N=300, SR 0.663) + frozen re-pin +
both Tier-B reach-floor guards (PASS: benign non-reach 0/30; adversary reach 0.042 m);
gitignores the 80MB raw episodes. Notes the box is now shared (user 40456786 on gpus 0,1).

+++ src/evasion_tax/attack/semantic_target.py
@@ -65,7 +65,13 @@ def build_semantic_target(
     prompt = _PROMPT_PREFIX.format(instruction=adv_instruction) + _PROMPT_TAIL
     inputs = processor(prompt, _to_pil(image))
     if hasattr(inputs, "to"):
-        inputs = inputs.to(device)
+        # OpenVLA's vision backbone runs in the model dtype (bf16); the processor emits
+        # float32 pixel_values, so cast the whole batch to model.dtype exactly as OpenVLA's
+        # own get_vla_action does (`.to(DEVICE, dtype=bf16)`). HF BatchFeature.to casts only
+        # the float tensors, leaving input_ids/attention_mask integer. Without this,
+        # model.generate raises "Input type (float) and bias type (BFloat16) should be the
+        # same" at the vision conv (box-verified 2026-07-10); gcg_openvla casts identically.
+        inputs = inputs.to(device, dtype=model.dtype)
     generated = model.generate(
         **inputs, max_new_tokens=_N_ACTION_TOKENS, do_sample=False
     )
